+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 0.1.0 (Unreleased)
 
 FEATURES:
+
+## 0.1.0-alpha1
+
+FEATURES:
+- New resource for managing integrations
+- New data-source for Apono connector